refactor: extract admin health projection
